Move addItemToCart and removeItemFromCart from InventoryPage to CartPage
