+++ test/specs/checkout2.e2e.js
@@ -21,7 +21,7 @@ describe('UserStory: Checkout', () => {
 
         console.log(`TestCase_2: Logged in`)
  
-        await InventoryPage.addItemToCart('#add-to-cart-sauce-labs-backpack');
+        await CartPage.addItemToCart('#add-to-cart-sauce-labs-backpack');
 
         console.log("TestCase_2: Added item to cart");
